+++ server/src/main/java/com/continuuity/loom/provisioner/plugin/ResourceService.java
@@ -19,17 +19,14 @@ import com.continuuity.http.BodyConsumer;
 import com.continuuity.http.HttpResponder;
 import com.continuuity.loom.account.Account;
 import com.continuuity.loom.common.conf.Configuration;
-import com.continuuity.loom.common.conf.Constants;
+import com.continuuity.loom.common.zookeeper.LockService;
 import com.continuuity.loom.common.zookeeper.lib.ZKInterProcessReentrantLock;
 import com.continuuity.loom.scheduler.task.MissingEntityException;
 import com.continuuity.loom.store.provisioner.PluginMetaStoreService;
 import com.continuuity.loom.store.provisioner.PluginMetaStoreView;
 import com.continuuity.loom.store.provisioner.PluginStore;
-import com.google.common.base.Joiner;
 import com.google.common.util.concurrent.AbstractIdleService;
 import com.google.inject.Inject;
-import org.apache.twill.zookeeper.ZKClient;
-import org.apache.twill.zookeeper.ZKClients;
 import org.jboss.netty.buffer.ChannelBuffer;
 import org.jboss.netty.handler.codec.http.HttpResponseStatus;
 import org.slf4j.Logger;
@@ -49,15 +46,15 @@ public class ResourceService extends AbstractIdleService {
   private final Configuration conf;
   private final PluginStore pluginStore;
   private final PluginMetaStoreService metaStoreService;
-  private final ZKClient zkClient;
+  private final LockService lockService;
 
   @Inject
   private ResourceService(PluginStore pluginStore, PluginMetaStoreService metaStoreService,
-                          ZKClient zkClient, Configuration conf) {
+                          LockService lockService, Configuration conf) {
     this.conf = conf;
     this.pluginStore = pluginStore;
     this.metaStoreService = metaStoreService;
-    this.zkClient = ZKClients.namespace(zkClient, Constants.PLUGIN_LOCK_NAMESPACE);
+    this.lockService = lockService;
   }
 
   /**
@@ -348,12 +345,7 @@ public class ResourceService extends AbstractIdleService {
   // locks are namespaced by tenant and resource type and name. for example,
   // /tenant1/automator/chef-solo/cookbooks/reactor
   private ZKInterProcessReentrantLock getLock(Account account, ResourceType type, String name) {
-    String path = Joiner.on('/')
-      .join(account.getTenantId(),
-            type.getPluginType().name().toLowerCase(),
-            type.getPluginName(),
-            type.getTypeName(),
-            name);
-    return new ZKInterProcessReentrantLock(zkClient, path);
+    return lockService.getResourceLock(account.getTenantId(), type.getPluginType().name().toLowerCase(),
+                                       type.getPluginName(), type.getTypeName(), name);
   }
 }