+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Continuuity Loom CHANGELOG
 
 Unreleased
 ----------
+- MongoDB support for clusters ( Issues: #43 #130 )
 - Ruby testing with rspec/rubocop ( Issues: #128 #132 #133 )
 - Nginx support for clusters ( Issues: #131 )
 - Sensu monitoring support ( Issues: #167 )
@@ -16,7 +17,9 @@ Unreleased
 - Provisioner Multi Tenancy ( Issues: #290 #296 )
 - Updated provider plugin using fog ( Issues: #288 #310 )
 - Queue Metrics ( Issues: #301 )
+- Settable max for clusters and nodes in a tenant ( Issues: #302 )
 - Allow non-root users for SSH ( Issues: #314 #320 )
+- PHP module support ( Issues: #317 )
 
 v0.9.7 (May 27, 2014)
 ---------------------